Move inputs too new folder
