+++ src/components/form/CreateAccount/inputs/ImageSubmit.tsx
@@ -2,7 +2,7 @@ import * as ImagePicker from 'expo-image-picker';
 import { Center, FormControl, Image, Pressable, Text } from 'native-base';
 import React, { FC, useState } from 'react';
 
-import { SavePictureSVG } from '../../svg';
+import { SavePictureSVG } from '../../../svg';
 
 type Props = {
   onImageChange: (imageUri: string) => void;
